Move model InvocationSerializationParams to invocations schema file
